% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{fitted.semiauto_fit}
\alias{fitted.semiauto_fit}
\title{Fitted values method for semiauto_fit}
\usage{
\method{fitted}{semiauto_fit}(object, ...)
}
\arguments{
\item{object}{A \code{semiauto_fit} object.}

\item{...}{Ignored.}
}
\description{
Fitted values method for semiauto_fit
}
