% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{predict.semiauto_fit}
\alias{predict.semiauto_fit}
\title{Predict method for semiauto_fit}
\usage{
\method{predict}{semiauto_fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{semiauto_fit} object.}

\item{newdata}{Optional data frame with new observations. If \code{NULL},
return fitted values on training data.}

\item{...}{Ignored.}
}
\description{
Predict method for semiauto_fit
}
