% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_check.R
\name{print.semiauto_check}
\alias{print.semiauto_check}
\title{Print Method for semiauto_check}
\usage{
\method{print}{semiauto_check}(x, ...)
}
\arguments{
\item{x}{An object of class \code{semiauto_check}.}

\item{...}{Additional arguments.}
}
\description{
Print Method for semiauto_check
}
