% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{print.semiauto_fit}
\alias{print.semiauto_fit}
\title{Print semiauto_fit Object}
\usage{
\method{print}{semiauto_fit}(x, ...)
}
\arguments{
\item{x}{Object from semiauto_fit().}

\item{...}{Additional arguments.}
}
\description{
Print semiauto_fit Object
}
