% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_missing.R
\name{print.semiauto_missing}
\alias{print.semiauto_missing}
\title{Print Method for semiauto_missing}
\usage{
\method{print}{semiauto_missing}(x, ...)
}
\arguments{
\item{x}{An object of class \code{semiauto_missing}.}

\item{...}{Additional arguments.}
}
\description{
Print Method for semiauto_missing
}
