% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_prepare.R
\name{print.semiauto_prepare}
\alias{print.semiauto_prepare}
\title{Print Method for semiauto_prepare}
\usage{
\method{print}{semiauto_prepare}(x, ...)
}
\arguments{
\item{x}{An object of class \code{semiauto_prepare}.}

\item{...}{Additional arguments.}
}
\description{
Print Method for semiauto_prepare
}
