% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{residuals.semiauto_fit}
\alias{residuals.semiauto_fit}
\title{Residuals method for semiauto_fit}
\usage{
\method{residuals}{semiauto_fit}(object, ...)
}
\arguments{
\item{object}{A \code{semiauto_fit} object.}

\item{...}{Ignored.}
}
\description{
Residuals method for semiauto_fit
}
