% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_check.R
\name{semiauto_check}
\alias{semiauto_check}
\title{Check Input Data for semiautoReg}
\usage{
semiauto_check(
  data,
  y,
  x = NULL,
  time = NULL,
  id = NULL,
  allow_missing = FALSE,
  min_n = 10,
  min_unique = 3,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing the response and predictor variables.}

\item{y}{A character string specifying the response variable name.}

\item{x}{A character vector specifying predictor variable names. If NULL,
all columns except y, time, and id will be used as predictors.}

\item{time}{Optional character string specifying the time variable.}

\item{id}{Optional character string specifying the subject or group ID variable.}

\item{allow_missing}{Logical. If FALSE, missing values are reported as a problem.}

\item{min_n}{Minimum required number of observations.}

\item{min_unique}{Minimum number of unique values required for the response variable.}

\item{verbose}{Logical. If TRUE, prints the checking result.}
}
\value{
An object of class \code{semiauto_check}.
}
\description{
This function checks whether the input data is suitable for semiparametric
regression modeling in the semiautoReg package.
}
\examples{
semiauto_check(mtcars, y = "mpg", x = c("wt", "hp", "qsec"))

}
