% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{semiauto_fit}
\alias{semiauto_fit}
\title{Fit Semiparametric Regression with AR Errors}
\usage{
semiauto_fit(
  prepared,
  spline_var,
  df = 5,
  degree = 3,
  lambda = 0.1,
  max_iter = 50,
  tol = 1e-06
)
}
\arguments{
\item{prepared}{An object returned by \code{semiauto_prepare()}. This argument is required.}

\item{spline_var}{Character. Name of the predictor to be modeled nonparametrically using spline basis. This argument is required.}

\item{df}{Numeric. Degrees of freedom for the spline basis. Default is 5.}

\item{degree}{Numeric. Degree of the spline basis. Default is 3.}

\item{lambda}{Numeric. SCAD penalty parameter. Default is 0.1.}

\item{max_iter}{Numeric. Maximum number of iterations. Default is 50.}

\item{tol}{Numeric. Convergence tolerance. Default is 1e-6.}
}
\value{
A list containing fitted model results.
}
\description{
Fits a semiparametric regression model using
B-spline basis expansion, SCAD penalty,
and autoregressive error correction.
}
\examples{
prep <- semiauto_prepare(
  data = mtcars,
  y = "mpg",
  x = c("wt", "hp", "qsec")
)

fit <- semiauto_fit(
  prepared = prep,
  spline_var = "wt"
)

fit

}
