% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_missing.R
\name{semiauto_missing}
\alias{semiauto_missing}
\title{Missing Value Summary for semiautoReg}
\usage{
semiauto_missing(data, y = NULL, x = NULL, verbose = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{y}{Optional character string specifying the response variable name.}

\item{x}{Optional character vector specifying predictor variable names.
If both y and x are NULL, all columns in the data will be checked.}

\item{verbose}{Logical. If TRUE, prints the missing value summary.}
}
\value{
An object of class \code{semiauto_missing}.
}
\description{
This function summarizes missing values in the selected variables.
It reports the number and percentage of missing values for each variable,
as well as the number of complete and incomplete rows.
}
\examples{
semiauto_missing(mtcars)
semiauto_missing(mtcars, y = "mpg", x = c("wt", "hp", "qsec"))

}
