% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_plot.R
\name{semiauto_plot}
\alias{semiauto_plot}
\title{Plot diagnostics for a semiauto_fit object (ggplot2)}
\usage{
semiauto_plot(x, which = 1:6, combine = FALSE, title = NULL, ...)
}
\arguments{
\item{x}{A\code{semiauto_fit} object.}

\item{which}{Integer vector. Which panels to draw. Default \code{1:6}.}

\item{combine}{Logical.
If TRUE, panels are combine using patchwork.
IF FALSE, a named list of ggplot objects is returned.
Default is FALSE.}

\item{title}{Character. Optional overall title.}

\item{...}{Ignored.}
}
\value{
A \code{patchwork} figure (if \code{combine=TRUE}) or a named list
of \code{ggplot} objects.
}
\description{
Produces up to 6 ggplot2 panels:
\enumerate{
\item Estimated nonlinear component functions
\item Residuals vs Fitted Values
\item ACF of residuals
\item Observed vs Fitted time series
\item Structure identification summary
\item AR coefficient estimates }
}
\examples{
\dontrun{
data(macro_turkey)

prep=semiauto_prepare(
 data = macro_turkey,
 y = "inflasi",
 x = c("kurs", "m3", "industri", "kredit", "brent"),
 time = "Tarih"
 )

 fit = semiauto_fit(prepared=prep)

 #Return diagnostic plots as a list
 plots = semiauto_plot(fit)

 names(plots)

 plots$nonlinear
 plots$resid_fitted
 plots$acf
 plots$obs_fitted
 plots$structure
 plots$ar_coefs

 #Combine all panels
 semiauto_plot(fit,combine=TRUE)
}

}
