% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{summary.semiauto_fit}
\alias{summary.semiauto_fit}
\title{Summary Method for semiauto_fit}
\usage{
\method{summary}{semiauto_fit}(object, ...)
}
\arguments{
\item{object}{A \code{semiauto_fit} object.}

\item{...}{Ignored}
}
\description{
Summary Method for semiauto_fit
}
