% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiauto_fit.R
\name{print.semiauto_fit}
\alias{print.semiauto_fit}
\title{Print method for semiauto_fit}
\usage{
\method{print}{semiauto_fit}(x, ...)
}
\arguments{
\item{x}{A \code{semiauto_fit} object.}

\item{...}{Ignored.}
}
\description{
Print method for semiauto_fit
}
